+++ src/BGS/ch/croctus.c
@@ -163,7 +163,7 @@ void chCroctus_update(Actor *this){
                 func_80324D2C(4.5f, COMUSIC_43_ENTER_LEVEL_GLITTER);
                 subaddie_set_state_with_direction(this, CHCROCTUS_STATE_5_FED, 0.79f, 1);
                 if (this->actorTypeSpecificField == 5) {
-                    timedFunc_set_1(0.9f, (GenFunction_1) func_80387E40, (s32) this->marker);
+                    timedFunc_set_1(0.9f, (GenFunction_1) func_80387E40, (intptr_t)this->marker);
                 } else {
                     func_80326310(this);
                 }
@@ -172,7 +172,7 @@ void chCroctus_update(Actor *this){
                     timedFunc_set_1(1.1f, (GenFunction_1)func_80387E00, reinterpret_cast(s32, bgs_D_803907B8[this->actorTypeSpecificField]));
                     gcStaticCamera_activate(D_803907B0[this->actorTypeSpecificField-1]);
                 } else {
-                    timedFunc_set_1(0.8f, (GenFunction_1)chCroctus_jiggySpawn, (s32) this->marker);
+                    timedFunc_set_1(0.8f, (GenFunction_1)chCroctus_jiggySpawn, (intptr_t)this->marker);
                 }
                 __spawnQueue_add_2((GenFunction_2)func_80387D18, reinterpret_cast(s32, this->marker), 0x46);
             }
