+++ src/core2/ch/jiggy.c
@@ -79,7 +79,7 @@ void destroyJiggy(Actor *this, s32 jiggyFlag, s32 timerRunningFlag, s32 cameraId
 
     progressDialog_setAndTriggerDialog_0(dialogFlag);
     timedFunc_set_4(0.6f, (GenFunction_4)spawnJiggyDestroyedEffects, (s32)this->position[0], (s32)this->position[1], (s32)this->position[2], switchPressedFlag);
-    timedFunc_set_2(0.6f, (GenFunction_2)playJiggyDestroyedSoundsAndReset, (s32)this->marker, resetSwitchFlag);
+    timedFunc_set_2(0.6f, (GenFunction_2)playJiggyDestroyedSoundsAndReset, (intptr_t)this->marker, resetSwitchFlag);
     timedFunc_set_0(1.0f, (GenFunction_0)func_802BE720);
     timedFunc_set_1(3.9f, (GenFunction_1)resetFlag, switchPressedFlag);
     mapSpecificFlags_set(jiggyFlag, TRUE);
